requirements.txt
discord
requests
python-dotenv
os-sys
